requirements.txt
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.5.0
scikit-surprise>=1.1.3
streamlit>=1.28.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
requests>=2.28.0
python-dotenv>=0.19.0 